requirements.txt
flask[async]==3.0.2
python-dotenv==1.0.1
pdfplumber==0.10.3
langchain==0.1.9
tiktoken
cohere==5.15.0
openai>=1.0.0
numpy
google-generativeai 